+++ src/docs/CHANGELOG.md
@@ -6,6 +6,8 @@ It favors sequences with lower number of method calls and cumulative execution t
 That is, `--input-selection=ORIENTEERING` is similar to `--input-selection=SMALL_TESTS`,
 but it considers cumulative execution time more important than test size.
 
+Randoop no longer builds under Java 8.  It still runs under Java 8.
+
 
 Version 4.3.2 (January 8, 2023)
 -------------------------------